requirements.txt
# Core dependencies
PyQt6>=6.5.0
qasync>=0.27.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# OpenAI
openai>=1.0.0

# Database
aiosqlite>=0.19.0

# Image processing
Pillow>=10.0.0

# MCP
mcp>=0.1.0

pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0